Add stub files and tests for first release
When run, the tests check for the correct response from the furniture
seller NPC, that the player cannot buy a furniture item without the
correct amount of money, and that the player can buy furniture with
enough money. The stub files contain just enough code to avoid errors in
the test files.