Fix item not existing in EntityManager
Adds the new class of item "furniture_item" to the corresponding XML and
EntityMap/EntityViewFactory files, as previously they were missing
causing issues with BuyFurnitureChatActionTest. I also changed the test
so it now should pass, as the check for the player's money had the wrong
response string in the assertion.