Reorganise packages and fix FurnitureSellerNPC
The FurnitureSellerNPC James Oakheart now appears in Deniran after I
fixed a problem in deniran.xml. Furthermore, his behaviour is fixed and
he now makes use of the BuyFurnitureChatAction that Ruben created. I
also removed the BuyFurniture class as it is now pointless, so I renamed
the BuyFurnitureTest to FurnitureSellerNPCTest and also updated the test
with the updated responses from James.